refactor(shared/ui): move traps out of modal component
