+++ client/src/shared/ui/trap/EscapeTrap.js
