+++ client/src/shared/ui/trap/FocusTrap.js
