+++ client/src/shared/ui/trap/KeyboardInteractionTrap.js
