+++ client/src/shared/ui/trap/index.js
@@ -1,0 +1,14 @@
+/**
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
+ * under one or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information regarding copyright
+ * ownership.
+ *
+ * Camunda licenses this file to you under the MIT; you may not use this file
+ * except in compliance with the MIT License.
+ */
+
+export { default as EscapeTrap } from './EscapeTrap';
+export { default as FocusTrap } from './FocusTrap';
+export { default as GlobalClickTrap } from './GlobalClickTrap';
+export { default as KeyboardInteractionTrap } from './KeyboardInteractionTrap';
